requirements.txt
mistralai
python-dotenv
openai
flask
